Allow regular expressions to help find blueprints
I needed to locate all of the 'paas-sps' blueprints.
I modified the extract-blueprint-names script to
locate blueprints based on regular expressions to
include and exclude.

For example this is what I ran:
```
ruby extract-blueprint-names.rb "paas-sps.*$" "(.*testapp.*)|(paas-sps.*-alarms)$"
```

Which gave me pretty much all of the blueprints that 'deploy'
for Provisioning.